+++ packages/orq-rc/src/models/operations/updatedatapoint.ts
@@ -6020,7 +6020,7 @@ export const UpdateDatapointResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-06T05:32:10.908Z",
+    "2025-10-06T08:24:22.146Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -6082,7 +6082,7 @@ export const UpdateDatapointResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-06T05:32:10.908Z"))
+  updated: z.date().default(() => new Date("2025-10-06T08:24:22.146Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {