+++ packages/orq-rc/RELEASES.md
@@ -4089,3 +4089,13 @@ Based on:
 - [typescript v3.14.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.8] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.8 - packages/orq-rc
+
+## 2025-10-06 08:26:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.634.2 (2.721.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.9] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.9] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.9 - packages/orq-rc